wacayang_ai_api/requirements.txt
Flask
gunicorn
tensorflow
numpy
pillow
firebase-admin